Uniquest Adapters: fix im_uid lost when userIdAsEids is empty array
