+++ libraries/uniquestUtils/uniquestUtils.js
@@ -1,9 +1,21 @@
-import { deepAccess } from '../../src/utils.js';
+import { deepAccess, getBidIdParameter } from '../../src/utils.js';
+import { tryAppendQueryString } from '../urlUtils/urlUtils.js';
 
-export function getImuid(request) {
-  const eids = request.userIdAsEids || deepAccess(request, 'ortb2.user.ext.eids') || [];
+export function buildQueryString(request, bidderRequest, paramKey) {
+  const eids = (request.userIdAsEids?.length ? request.userIdAsEids : deepAccess(request, 'ortb2.user.ext.eids')) || [];
   const imuidEid = eids.find(eid => eid.source === 'intimatemerger.com');
-  return deepAccess(imuidEid, 'uids.0.id');
+  const imuid = deepAccess(imuidEid, 'uids.0.id');
+  const widths = request.sizes.map(size => size[0]).join(',');
+  const heights = request.sizes.map(size => size[1]).join(',');
+
+  let queryString = '';
+  queryString = tryAppendQueryString(queryString, 'bid', request.bidId);
+  queryString = tryAppendQueryString(queryString, paramKey, getBidIdParameter(paramKey, request.params));
+  queryString = tryAppendQueryString(queryString, 'widths', widths);
+  queryString = tryAppendQueryString(queryString, 'heights', heights);
+  queryString = tryAppendQueryString(queryString, 'timeout', bidderRequest.timeout);
+  queryString = tryAppendQueryString(queryString, 'im_uid', imuid);
+  return queryString;
 }
 
 export function interpretResponse (serverResponse) {
