+++ modules/uniquestBidAdapter.js
@@ -1,8 +1,6 @@
-import { getBidIdParameter } from '../src/utils.js';
 import { registerBidder } from '../src/adapters/bidderFactory.js';
 import { BANNER } from '../src/mediaTypes.js';
-import { tryAppendQueryString } from '../libraries/urlUtils/urlUtils.js';
-import { interpretResponse, getImuid } from '../libraries/uniquestUtils/uniquestUtils.js';
+import { interpretResponse, buildQueryString } from '../libraries/uniquestUtils/uniquestUtils.js';
 
 /**
  * @typedef {import('../src/adapters/bidderFactory').Bid} Bid
@@ -35,33 +33,11 @@ export const spec = {
    * @return ServerRequest Info describing the request to the server.
    */
   buildRequests: function (validBidRequests, bidderRequest) {
-    const bidRequests = [];
-
-    for (let i = 0; i < validBidRequests.length; i++) {
-      let queryString = '';
-      const request = validBidRequests[i];
-
-      const bid = request.bidId;
-      const sid = getBidIdParameter('sid', request.params);
-      const widths = request.sizes.map(size => size[0]).join(',');
-      const heights = request.sizes.map(size => size[1]).join(',');
-      const timeout = bidderRequest.timeout;
-      const imuid = getImuid(request);
-
-      queryString = tryAppendQueryString(queryString, 'bid', bid);
-      queryString = tryAppendQueryString(queryString, 'sid', sid);
-      queryString = tryAppendQueryString(queryString, 'widths', widths);
-      queryString = tryAppendQueryString(queryString, 'heights', heights);
-      queryString = tryAppendQueryString(queryString, 'timeout', timeout);
-      queryString = tryAppendQueryString(queryString, 'im_uid', imuid);
-
-      bidRequests.push({
-        method: 'GET',
-        url: ENDPOINT,
-        data: queryString,
-      });
-    }
-    return bidRequests;
+    return validBidRequests.map(request => ({
+      method: 'GET',
+      url: ENDPOINT,
+      data: buildQueryString(request, bidderRequest, 'sid'),
+    }));
   },
   interpretResponse: interpretResponse,
 };
