+++ test/spec/modules/uniquest_widgetBidAdapter_spec.js
@@ -85,6 +85,25 @@ describe('uniquest_widgetBidAdapter', function () {
       expect(requests[0].data).to.include('im_uid=test_imuid_123');
     });
 
+    it('includes im_uid from ortb2 when userIdAsEids is an empty array', function () {
+      const bidsWithEid = [{
+        ...bids[0],
+        userIdAsEids: [],
+        ortb2: {
+          user: {
+            ext: {
+              eids: [{
+                source: 'intimatemerger.com',
+                uids: [{ id: 'test_imuid_123' }],
+              }],
+            },
+          },
+        },
+      }];
+      const requests = spec.buildRequests(bidsWithEid, bidderRequest);
+      expect(requests[0].data).to.include('im_uid=test_imuid_123');
+    });
+
     it('does not include im_uid in query string when imuid EID is absent', function () {
       const requests = spec.buildRequests(bids, bidderRequest);
       expect(requests[0].data).to.not.include('im_uid');
